requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.2.0
amqp==5.1.1
annotated-types==0.6.0
anyio==4.2.0
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
b2sdk==1.30.1
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.34.28
botocore==1.34.28
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorgram.py==1.2.0
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.1.5
django-ajax==0.3.0
django-anymail==10.0
django-background-tasks==1.2.5
django-browser-reload==1.6.0
django-chartjs==2.3.0
django-ckeditor==6.5.1
django-compat==1.0.15
django-etc==1.4.0
django-hitcount==1.3.5
django-htmx==1.17.2
django-import-export==3.3.1
django-js-asset==2.0.0
django-model-utils==4.4.0
django-notifications-hq==1.8.3
django-pdfkit==0.3.1
django-storages==1.14.2
django-timezone-field==5.1
django-wkhtmltopdf==3.4.0
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.41
gTTS==2.5.0
gunicorn==21.2.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonfield==3.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.1
libretranslatepy==2.1.1
logfury==1.0.1
lxml==4.9.2
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==2.1.4
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.3
odfpy==1.4.1
openai==0.28.0
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pandas==2.2.0
pdfkit==0.6.0
Pillow==9.4.0
prompt-toolkit==3.0.39
protobuf==4.25.2
psycopg2-binary==2.9.9
pyarrow==15.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyPDF3==1.0.6
pypng==0.20220715.0
pyquery==2.0.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
readtime==3.0.0
redis==4.6.0
referencing==0.32.1
reportlab==4.0.9
requests==2.28.2
rich==13.7.0
rpds-py==0.17.1
s3transfer==0.10.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.3
svglib==1.5.1
swapper==1.3.0
tablib==3.5.0
tenacity==8.2.3
threadpoolctl==3.3.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.64.1
translate==3.6.1
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.2
urllib3==1.26.14
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
xhtml2pdf==0.2.8
yarl==1.9.2
zipp==3.17.0
